% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimTimeVar.R
\name{add_one_categorical}
\alias{add_one_categorical}
\title{Generate linear predictor from logistic model}
\usage{
add_one_categorical(.d, n, obs, cat.parameters)
}
\arguments{
\item{.d}{The dataset to which to add the categorical variable.}

\item{n}{The number of clusters.}

\item{obs}{The number of observations per cluster.}

\item{cat.parameters}{A dataframe of parameters for generating the categorical variable. See Details.}
}
\description{
An internal function not intended for the user. Given a dataset and multinomial regression parameters,
generates a categorical variable and adds it to the dataset.
}
\examples{
# mini dataset with 3 observations per person
data = data.frame( male = rep( rbinom(n=10, size=1, prob=0.5), each=3 ) )
add_one_categorical( data, 10, 3, cat.params)
}
