% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimTimeVar.R
\name{add_time_function_vars}
\alias{add_time_function_vars}
\title{Creates linear time-function variables}
\usage{
add_time_function_vars(d4, obs, parameters)
}
\arguments{
\item{d4}{The dataframe to which to add the time-function variables.}

\item{obs}{The number of observations per cluster.}

\item{parameters}{The parameters matrix.}
}
\description{
Given variable-specific slopes and intercepts for a cluster, creates continuous variables that
increase or 
decrease linearly in time (with normal error with standard deviation \code{error.SD}) and
adds them to the dataframe.
}
\details{
See \code{make_one_dataset} for additional information.
}
