% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimTimeVar.R
\name{BN.rBound}
\alias{BN.rBound}
\title{Maximum correlation between binary and normal random variables}
\usage{
BN.rBound(p)
}
\arguments{
\item{p}{Parameter of Bernoulli random variable.}
}
\description{
Given parameter \code{p} for a Bernoulli random variable, returns its maximum possible
correlation with an arbitrary normal random variable. Used to adjust correlation matrices
whose entries are not theoretically possible.
}
\examples{
# find the largest possible correlation between a normal
#  variable and a binary with parameter 0.1
BN.rBound(0.1)
}
