% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimTimeVar.R
\docType{data}
\name{cat.params}
\alias{cat.params}
\title{An example dataframe for categorical variable parameters}
\format{An object of class \code{data.frame} with 5 rows and 3 columns.}
\usage{
cat.params
}
\description{
An example of how to set up the categorical variable parameters dataframe.
}
\keyword{datasets}
