% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimTimeVar.R
\name{closest}
\alias{closest}
\title{Return closest value}
\usage{
closest(x, candidates)
}
\arguments{
\item{x}{The number to be compared to the permitted values.}

\item{candidates}{A vector of permitted values.}
}
\description{
An internal function not intended for the user. Given a number \code{x} and vector of
 permitted values, 
returns the closest permitted value to \code{x} (in absolute value).
}
\examples{
closest( x = 5, candidates = c(-3, 8, 25) )
}
