% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimTimeVar.R
\name{expand_matrix}
\alias{expand_matrix}
\title{Longitudinally expand a matrix of single observations by cluster}
\usage{
expand_matrix(.matrix, .obs)
}
\arguments{
\item{.matrix}{The matrix of observations to be expanded.}

\item{.obs}{The number of observations to generate per cluster.}
}
\description{
An internal function not intended for the user. Given a matrix of single observations 
for a cluster, repeats each cluster's entry in each \code{.obs} times.
}
\examples{
mat = matrix( seq(1:10), nrow=2, byrow=FALSE)
expand_matrix(mat, 4)
}
