% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimTimeVar.R
\name{has_drug_suffix}
\alias{has_drug_suffix}
\title{Checks whether string has "_s" suffix}
\usage{
has_drug_suffix(var.name)
}
\arguments{
\item{var.name}{The string to be checked}
}
\description{
An internal function not intended for the user.
}
\examples{
has_drug_suffix("myvariable_s")
has_drug_suffix("myvariable")
}
