% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimTimeVar.R
\name{make_one_linear_pred}
\alias{make_one_linear_pred}
\title{Generate linear predictor from logistic model}
\usage{
make_one_linear_pred(m, data)
}
\arguments{
\item{m}{Part of the parameter matrix for the linear predictor for a single variable.}

\item{data}{The dataframe from which to generate.}
}
\description{
An internal function not intended for the user. Given a matrix of regression parameters
and a dataset, returns the linear predictor based on the given dataset.
}
\examples{
# take part of parameters matrix corresponding to single level of categorical
#  variable
m = cat.params[ cat.params$level == "black", ]
data = data.frame( male = rbinom(n=10, size=1, prob=0.5) )
make_one_linear_pred(m, data)
}
