% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimTimeVar.R
\name{mod.jointly.generate.binary.normal}
\alias{mod.jointly.generate.binary.normal}
\title{Return closest value}
\usage{
mod.jointly.generate.binary.normal(no.rows, no.bin, no.nor, prop.vec.bin,
  mean.vec.nor, var.nor, corr.vec, adjust.corrs = TRUE)
}
\arguments{
\item{no.rows}{Number of rows}

\item{no.bin}{Number of binary variables}

\item{no.nor}{Number of normal variables}

\item{prop.vec.bin}{Vector of parameters for binary variables}

\item{mean.vec.nor}{Vector of means for binary variables}

\item{var.nor}{Vector of variances for binary variables}

\item{corr.vec}{Vector of correlations}

\item{adjust.corrs}{Boolean indicating whether theoretically impossible correlations between 
a binary and a normal variable should be adjusted to their closest theoretically possible value.}
}
\description{
An internal function not intended for the user. Simulates correlated normal and binary
variables based on the algorithm of Demirtas and Doganay (2012). See references for
further information.
}
\references{
Demirtas, H., & Doganay, B. (2012). Simultaneous generation of binary and
normal data with specified marginal and association structures. Journal of
Biopharmaceutical Statistics, 22(2), 223-236.
}
