% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimTimeVar.R
\docType{data}
\name{params}
\alias{params}
\title{An example parameters dataframe}
\format{An object of class \code{data.frame} with 12 rows and 8 columns.}
\usage{
params
}
\description{
An example of how to set up the parameters dataframe.
}
\keyword{datasets}
