% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimTimeVar.R
\docType{data}
\name{pcor}
\alias{pcor}
\title{An example across-cluster correlation dataframe}
\format{An object of class \code{data.frame} with 9 rows and 9 columns.}
\usage{
pcor
}
\description{
An example of how to set up the across-cluster correlation dataframe.
}
\keyword{datasets}
