% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimTimeVar.R
\name{proportionize}
\alias{proportionize}
\title{Turn a number into a valid proportion}
\usage{
proportionize(x, zero = 1e-05, one = 0.999)
}
\arguments{
\item{x}{The number to be turned into a proportion.}

\item{zero}{A very small number that is just larger than 0.}

\item{one}{A number that is just smaller than 1.}
}
\description{
An internal function not intended for the user. Turns an arbitrary number into a valid proportion
by setting the number equal to the closest value in [0,1].
}
\examples{
proportionize(-0.03)
proportionize(1.2)
proportionize(.63)
}
