% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimTimeVar.R
\name{upper_tri_vec}
\alias{upper_tri_vec}
\title{Turn symmetric matrix into vector}
\usage{
upper_tri_vec(m)
}
\arguments{
\item{m}{Matrix}
}
\description{
An internal function not intended for the user. Turns a matrix into a vector of the upper-triangular
elements (arranged by row).
}
\examples{
# make a simple correlation matrix
x = rnorm(10); y = rnorm(10); z = rnorm(10)
mat = cor( data.frame(x,y,z) )

# turn into into vector
upper_tri_vec(mat)
}
