% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimTimeVar.R
\docType{data}
\name{wcor}
\alias{wcor}
\title{An example within-cluster correlation dataframe}
\format{An object of class \code{data.frame} with 6 rows and 6 columns.}
\usage{
wcor
}
\description{
An example of how to set up the within-cluster correlation dataframe.
}
\keyword{datasets}
